+++ apps/desktop-wallet/src/api/apiDataHooks/useAddressesAlphBalances.ts
@@ -20,7 +20,7 @@ import { AddressHash } from '@alephium/shared'
 import { useQueries, UseQueryResult } from '@tanstack/react-query'
 
 import { addressAlphBalancesQuery, AddressAlphBalancesQueryFnData } from '@/api/addressQueries'
-import { useAddressesLastTransactionHashes } from '@/api/addressTransactionsDataHooks'
+import useAddressesLastTransactionHashes from '@/api/apiDataHooks/useAddressesLastTransactionHashes'
 import { combineIsLoading } from '@/api/apiDataHooks/utils'
 import { useAppSelector } from '@/hooks/redux'
 import { DisplayBalances } from '@/types/tokens'
