+++ apps/desktop-wallet/src/api/apiDataHooks/useAddressesTokensBalances.ts
@@ -22,7 +22,7 @@ import { useQueries } from '@tanstack/react-query'
 import { useMemo } from 'react'
 
 import { addressTokensBalanceQuery, AddressTokensBalancesQueryFnData } from '@/api/addressQueries'
-import { useAddressesLastTransactionHashes } from '@/api/addressTransactionsDataHooks'
+import useAddressesLastTransactionHashes from '@/api/apiDataHooks/useAddressesLastTransactionHashes'
 import useAddressesAlphBalances, { AddressesAlphBalances } from '@/api/apiDataHooks/useAddressesAlphBalances'
 import { mapCombine } from '@/api/apiDataHooks/utils'
 import { useAppSelector } from '@/hooks/redux'
